core_platform/requirements.txt
# XRPL DEX Platform Requirements
# Core dependencies for the complete DeFi ecosystem

# XRPL Integration
xrpl-py>=2.0.0
xrpl-hooks>=1.0.0

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Security and Cryptography
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Async and Concurrency
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0
tenacity>=8.2.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Optional: Advanced Features
# tensorflow>=2.15.0  # For AI-powered features
# torch>=2.1.0        # Alternative ML framework
# plotly>=5.17.0      # For advanced charts
# dash>=2.14.0        # For web dashboards
